requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests==2.32.5
speedtest-cli==2.1.3
urllib3==2.5.0
